**Important: Change in Security Implementation**
As of this commit, we have decided to withdraw the security measures previously implemented in this software. Instead of focusing on security, we have chosen to incorporate a license that will govern the use, modification and distribution of the code.

Please note that when making this transition, it is the user's responsibility to implement appropriate security measures as necessary for their deployment environment.

We appreciate your understanding and encourage you to review the new license [license name] included in this repository to understand the terms and conditions of use.

For any questions or concerns, please do not hesitate to contact us.